[Step3] refactor: 예약관련 파일들 reservation 패키지 이동
