+++ src/main/java/nextstep/Main.java
@@ -1,5 +1,7 @@
 package nextstep;
 
+import nextstep.reservation.*;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.Scanner;
