+++ src/main/java/nextstep/reservation/ReservationDeleteRequest.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
