+++ src/test/java/nextstep/reservation/ReservationControllerTest.java
@@ -1,8 +1,10 @@
-package nextstep;
+package nextstep.reservation;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import nextstep.reservation.ReservationCreateRequest;
+import nextstep.reservation.ReservationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
